Rename components/DataViewDataSetPanel.vue to components/index/_shared/DataViewDataSetPanel.vue
