+++ components/index/CardsReference/ConfirmedCasesAttributes/DataTable.vue
@@ -88,8 +88,8 @@ import { mdiAlert } from '@mdi/js'
 import Vue from 'vue'
 import ScaleLoader from 'vue-spinner/src/ScaleLoader.vue'
 
-import DataViewDataSetPanel from '@/components/DataViewDataSetPanel.vue'
 import DataView from '@/components/index/_shared/DataView.vue'
+import DataViewDataSetPanel from '@/components/index/_shared/DataViewDataSetPanel.vue'
 import OpenDataLink from '@/components/index/_shared/OpenDataLink.vue'
 
 export default Vue.extend({
